+++ batch-console-api/src/test/java/io/github/pinpols/batch/console/domain/rbac/support/ConsoleTenantGuardTest.java
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.when;
 
 import io.github.pinpols.batch.common.enums.ResultCode;
 import io.github.pinpols.batch.common.exception.BizException;
+import io.github.pinpols.batch.console.shared.security.ConsolePrincipal;
 import io.github.pinpols.batch.console.support.web.ConsoleRequestMetadataResolver;
 import java.util.Set;
 import org.junit.jupiter.api.AfterEach;